nixos/dockerPreloader: preload docker images
This module permits to preload Docker image in a VM in order to reduce
file copies. This has to be only used in testing environments, when
the test requires several Docker images.

The idea is to populate the /var/lib/docker directory by mounting qcow
files (we uses qcow file to avoid permission issues) that contain images.

For each image specified in
config.virtualisation.dockerPreloader.images:
1. The image is loaded by Docker in a VM
2. The resulting /var/lib/docker is written to a QCOW file

This set of QCOW files can then be used to populate the
/var/lib/docker:
1/ Each QCOW is mounted in the VM
2/ Symlink are created from these mount points to /var/lib/docker
3/ A /var/lib/docker/image/overlay2/repositories.json file is generated
4/ The docker daemon is started.